pyproject.toml
[tool.poetry]
name = 'schemarecomb'
version = '0.0.4'
description = 'Design recombinant protein libraries for Golden Gate Assembly.'
authors = ['Bennett Bremer <bennettjamesbremer@gmail.com>']

[tool.poetry.dependencies]
python = '~3.9.1'
biopython = '^1.79'
scipy = '^1.6.3'

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
mypy = "^0.910"
pytest = '^6.2.4'
pytest-mock = '^3.6.1'
sphinx = '^4.0.1'
sphinx-autodoc-typehints = '^1.12.0'

[tool.poetry.urls]
"Download" = "https://pypi.org/project/schemarecomb"
"Source Code" = "https://github.com/RomeroLab/schemarecomb"
"Documentation" = "https://schemarecomb.readthedocs.io/en/latest"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
doctest_optionflags = "ELLIPSIS"
testpaths = [
    "tests/unit/",
    "src/schemarecomb/libraries.py",
    "src/schemarecomb/pdb_structure.py",
    "src/schemarecomb/parent_alignment.py",
    "README.rst"
]

# [tool.poetry.scripts]
# schemarecomb = {reference = "bin/schemarecomb", type="file"}

[build-system]
requires = ["poetry-core>=1.1.0a5"]
build-backend = "poetry.core.masonry.api"
